Change thing types to String, reuse HTTP client, add headers and comments for code checks, format comments

+++ src/main/java/org/openhab/binding/venstarthermostat/model/VenstarInfoData.java
@@ -1,11 +1,23 @@
+/**
+ * Copyright (c) 2010-2018 by the respective copyright holders.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
 package org.openhab.binding.venstarthermostat.model;
 
+/**
+ * The {@link VenstarInfoData} class defines the vendor info data from a thermostat
+ *
+ * @author William Welliver - Initial contribution
+ */
 public class VenstarInfoData {
     double cooltemp;
     double heattemp;
 
-    double state;
-    double mode;
+    int state;
+    int mode;
 
     public double getCooltemp() {
         return cooltemp;
@@ -23,19 +35,19 @@ public class VenstarInfoData {
         this.heattemp = heattemp;
     }
 
-    public double getState() {
+    public int getState() {
         return state;
     }
 
-    public void setState(double state) {
+    public void setState(int state) {
         this.state = state;
     }
 
-    public double getMode() {
+    public int getMode() {
         return mode;
     }
 
-    public void setMode(double mode) {
+    public void setMode(int mode) {
         this.mode = mode;
     }
